+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Fixed
 
+- Fixed Equipment-Editor not showing the current synced values, but the cached ones 
 - Fix serverConVars indexing 0 at tables (could cause unpredictable issues)
 
 ## [v0.11.2b](https://github.com/TTT-2/TTT2/tree/v0.11.2b) (2021-11-17)